requirements.txt
smolagents>=0.1.0
huggingface-hub>=0.19.0
transformers>=4.36.0
torch>=2.1.0
python-dotenv>=1.0.0 

